+++ .storybook/preview.ts
@@ -1,6 +1,31 @@
 import type { Preview } from "@storybook/react-webpack5";
 
 const preview: Preview = {
+  args: {
+    className: "ood-primary",
+  },
+  argTypes: {
+    children: {
+      table: {
+        disable: true,
+      },
+    },
+    id: {
+      table: {
+        disable: true,
+      },
+    },
+    onClick: {
+      table: {
+        disable: true,
+      },
+    },
+    style: {
+      table: {
+        disable: true,
+      },
+    },
+  },
   parameters: {
     controls: {
       matchers: {