+++ src/index.ts
@@ -1,9 +1,14 @@
+export { Button } from "./components/Button";
 export { Draggable } from "./components/Draggable";
 export { Expandable } from "./components/Expandable";
 export type {
   IChildlessComponentProps,
   IComponentProps,
 } from "./components/IComponent";
+export { Infobox } from "./components/infobox/Infobox";
+export { InfoboxRow } from "./components/infobox/InfoboxRow";
+export { InfoboxSection } from "./components/infobox/InfoboxSection";
+export { InfoboxTitle } from "./components/infobox/InfoboxTitle";
 export { NavBar } from "./components/NavBar";
 export { NavDropdown } from "./components/NavDropdown";
 export { NavLink } from "./components/NavLink";