+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.2 (2025-10-01)
+
+### Features
+
+- Add `Button` component.
+- Add `Infobox` components.
+
+### Bug Fixes
+
+- Allow `Draggable` consumers to be aware of dragging state.
+
 ## 0.0.1 (2025-09-08)
 
 ### Features